+++ LibraryComponents/Classes/Core/AndesCard/BodyPadding/AndesCardBodyPaddingMedium.swift
@@ -1,0 +1,12 @@
+//
+//  AndesCardBodyPaddingMedium.swift
+//  AndesUI
+//
+//  Created by Federico Gabriel Nosmor on 19/11/2020.
+//
+
+import Foundation
+
+struct AndesCardBodyPaddingMedium: AndesCardBodyPaddingProtocol {
+    let bodyPaddingSize: UInt = 24
+}