+++ LibraryComponents/Classes/Core/AndesCard/BodyPadding/AndesCardBodyPaddingNone.swift
@@ -1,0 +1,12 @@
+//
+//  AndesCardBodyPaddingNone.swift
+//  AndesUI
+//
+//  Created by Federico Gabriel Nosmor on 19/11/2020.
+//
+
+import Foundation
+
+struct AndesCardBodyPaddingNone: AndesCardBodyPaddingProtocol {
+    let bodyPaddingSize: UInt = 0
+}