+++ LibraryComponents/Classes/Core/AndesCard/Config/AndesCardViewConfig.swift
@@ -13,6 +13,7 @@ internal struct AndesCardViewConfig {
     let pipeColor: UIColor
     let borderColor: UIColor?
     let padding: UInt
+    let bodyPadding: UInt
     let titlePadding: UInt
     let titleHeight: UInt
     let shadow: AndesCardShadowConfig?
@@ -22,11 +23,12 @@ internal struct AndesCardViewConfig {
     var linkText: String?
     let cardView: UIView
 
-    init(backgroundColor: UIColor, pipeColor: UIColor, borderColor: UIColor?, padding: UInt, titlePadding: UInt, titleHeight: UInt, shadow: AndesCardShadowConfig?, hasCardAction: Bool, titleText: String?, linkText: String?, cardView: UIView) {
+    init(backgroundColor: UIColor, pipeColor: UIColor, borderColor: UIColor?, padding: UInt, bodyPadding: UInt, titlePadding: UInt, titleHeight: UInt, shadow: AndesCardShadowConfig?, hasCardAction: Bool, titleText: String?, linkText: String?, cardView: UIView) {
         self.backgroundColor = backgroundColor
         self.pipeColor = pipeColor
         self.borderColor = borderColor
         self.padding = padding
+        self.bodyPadding = bodyPadding
         self.titlePadding = titlePadding
         self.titleHeight = titleHeight
         self.shadow = shadow
@@ -41,6 +43,7 @@ internal struct AndesCardViewConfig {
         self.pipeColor = .clear
         self.borderColor = nil
         self.padding = 0
+        self.bodyPadding = 0
         self.titlePadding = 16
         self.titleHeight = 40
         self.shadow = nil