+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unpublished
+### 🛠 Bug fixes
+- AndestextField: Fix icon component right and left interaction | Authors: [@joalonsopint](https://github.com/joalonsopint)
+
 # v3.16.0
 ### 🚀 Features
 - AndesCard: New property to make the padding independent in the body | Authors: [@federiconosmor](https://github.com/federiconosmor)