+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-# unpublised
+# Unpublished
 ### 🚀 Features
+- AndesProgressIndicatorIndeterminate: New component AndesProgressIndicator| Authors: [@javff](https://github.com/javff)
 - AndesDropdown  component  | Authors: [@joalonsopint](https://github.com/joalonsopint)
+
+### 🛠 Bug fixes
+- AndestextField: Fix icon component right and left interaction | Authors: [@joalonsopint](https://github.com/joalonsopint)
 
 # v3.16.0
 ### 🚀 Features